Fix : 백준 공고별 전체 및 1대1 비교 퍼센테이지 수정

+++ backend/src/main/java/com/ssafy/backend/domain/analysis/dto/BojLanguagePassCount.java
@@ -7,10 +7,10 @@ import lombok.Getter;
 @Getter
 public class BojLanguagePassCount {
 	private Long languageId;
-	private double totalPassCount;
+	private Integer totalPassCount;
 
 	@QueryProjection
-	public BojLanguagePassCount(Long languageId, Double totalPassCount) {
+	public BojLanguagePassCount(Long languageId, Integer totalPassCount) {
 		this.languageId = languageId;
 		this.totalPassCount = totalPassCount;
 	}
